Revert "5月30日更新"
This reverts commit 7a91dc59305ef0307d26cd49c943097510ee1aa1.

+++ Demo1_Jewel/Jewel.h
@@ -8,12 +8,11 @@ class Jewel :
 {
 public:
 	//宝石的位置
-	float pos_row;
-	float pos_col;
+	int pos_row;
+	int pos_col;
 	//宝石的控制按钮
 	Button* jewel_btn;
 	Sprite* border;
-	//std::vector<std::vector<Jewel*>> map;
 	//宝石是否被选中
 	bool isSelected;
 	Jewel();
